coloring_book/requirements_video.txt
moviepy==1.0.3
elevenlabs==0.2.27
python-dotenv==1.0.1
Pillow==10.2.0
